Add file
This is a game made by Python3. It has no GUI, the user needs to play it via terminal. In this game, the code generates a random digit from 1-9 every single round. The player needs to guess the correct secret digit in only 3 chances. If you can guess within 3 chances, you win! 🏆